+++ src/include/hooks.h
@@ -5,6 +5,6 @@
 
 ssize_t hook_input (char*, size_t);
 ssize_t hook_output (const char*, size_t);
-ssize_t hook_seek (off_t);
+off_t hook_seek (off_t);
 
 #endif /* _LIBDESOCK_HOOKS_H */
